Calculate Quarterly Days in AR

Calculate Quarterly Days in AR

Usage

dar_quarter(df, date, gct, earb, dart = 35)

Arguments

df

<data.frame> containing at least three columns: a date column, a gross charges column and an ending AR column

date

column containing a date within the month that Days in AR is to be calculated

gct

column containing a month's total Gross Charges

earb

column containing a month's Ending AR balance

dart

Days in AR target, default is 35 days

Value

a tibble

Examples

dar_quarter(
   dar_example(),
   date,
   gct,
   earb,
   dart = 40)
#> # A tibble: 4 × 14
#>   date        nqtr  ndip gct_qtr    earb earb_trg  earb_dc earb_pct   adc   dar
#>   <date>     <int> <int>   <dbl>   <dbl>    <dbl>    <dbl>    <dbl> <dbl> <dbl>
#> 1 2024-03-01     1    91 822369. 253977.  361481. -107504.   -42.3  9037.  28.1
#> 2 2024-06-01     2    91 441142. 203460.  193909.    9552.     4.69 4848.  42.0
#> 3 2024-09-01     3    92 457846. 179348.  199063.  -19716.   -11.0  4977.  36.0
#> 4 2024-12-01     4    92 484305. 199849.  210567.  -10718.    -5.36 5264.  38.0
#> # ℹ 4 more variables: pass <lgl>, actual <dbl>, ideal <dbl>, radiff <dbl>